API Documentation

The API for the package

Standard functions

There is a range of predefined functions implemented as a struct that holds tuple of parameters

Functions

AlgebraPDF.noparsnormf — Function

noparsnormf(d::AbstractPDF; p=pars(d))

Returns a single-argument lambda-function with parameters fixed to p and normalization computed.

AlgebraPDF.nt — Function

nt(s::Symbol, v = 0.0)

Creates a named tuple (s=v,) where s is a provided symbol, and v is the value.

Macros

AlgebraPDF.@makepdftype — Macro

@makepdftype MyPDF(x;p) = unnormdensity(x, p.a, p.b)

Expected form of the expression is f(x;p) on the left

AlgebraPDF.@makefuntype — Macro

@makefuntype MyPDF(x;p) = unnormdensity(x, p.a, p.b)

Expected form of the expression is `f(x;p)` on the left

Other Functions

AlgebraPDF.updatevalueorflag — Function

updatevalueorflag(p::FlaggedNamedTuple, s::Symbol, isfree::Bool, v=getproperty(pars(p),s))

Implementation of the main update method for FlaggedNamedTuple parameters.